rename discover pages
