+++ .github/workflows/ci.yml
@@ -97,3 +97,25 @@ jobs:
         run: npm run compile
       - name: Check no changes were made
         run: git diff --exit-code
+  codegen:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Setup Node.js
+        uses: actions/setup-node@v2
+        with:
+          node-version: 16.15.0
+      - uses: actions/cache@v2
+        with:
+          path: ~/.npm
+          key: ${{ runner.os }}-node-${{ hashFiles('**/package-lock.json') }}
+          restore-keys: |
+            ${{ runner.os }}-node-
+      - name: Install dependencies
+        run: npm ci
+      - name: Purge
+        run: npm run graphql:purge
+      - name: Codegen
+        run: npm run graphql:codegen
+      - name: Check no changes were made
+        run: git diff --exit-code
