+++ src/containers/discover/index.slider.module.scss
