+++ src/containers/discover/index.slider.spec.tsx
@@ -6,7 +6,7 @@ import { __eventHandlers, __runHandlers, __swiper } from '~lib/swiper';
 import useElementWidth from '~src/lib/hooks/useElementWidth';
 import { buildRenderer } from '~src/lib/test/buildRenderer';
 
-import Slider, { GRID_GAP, MIN_CARD_WIDTH } from './discover.slider';
+import Slider, { GRID_GAP, MIN_CARD_WIDTH } from './index.slider';
 
 jest.mock('~lib/hooks/useElementWidth');
 
@@ -29,7 +29,7 @@ export const mockWidth = (cardsPerPage: number) => {
 
 const getSlides = () => {
 	const container = screen.getByTestId('swiper');
-	
+
 	return within(container).getAllByTestId('swiper-slide');
 }
