+++ src/containers/discover/index.slider.tsx
